% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultPlotParams.R
\name{getDefaultPlotParams}
\alias{getDefaultPlotParams}
\title{getDefaultPlotParams}
\usage{
getDefaultPlotParams(show.introns = FALSE, compress.introns = TRUE,
  proportional.introns = TRUE)
}
\description{
getDefaultPlotParams
}
\note{
: TODO documentation
}
