% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpAddBaseNumbers.R
\name{gpAddBaseNumbers}
\alias{gpAddBaseNumbers}
\title{gpAddBaseNumbers}
\usage{
gpAddBaseNumbers(genoplot, label.exon.ends=TRUE, label.every.x=FALSE,
                           add.horizontal.line=FALSE, add.minor.ticks=TRUE,
                           custom.labels=NULL, ticks.in.introns=TRUE,
                           position="below.gene", inverted=FALSE,
                           add.units=TRUE, digits=2, shorten.labels=TRUE, cDNA=FALSE,
                           cex=0.6, label.col=NULL,
                           tick.dist=500, tick.len=0.2, tick.col=NULL,
                           minor.tick.dist=100, minor.tick.len=0.5, minor.tick.col=NULL,
                           r0=0, r1=1, data.panel=1,
                           col="black", pos=NULL, ...)
}
\arguments{
\item{genoplot}{(GEnoPlot object) The GenoPlot object returned by the call to \code{\link{plotGene}}}

\item{label.exon.ends}{(defaults to TRUE)}

\item{label.every.x}{(defaults to FALSE)}

\item{add.horizontal.line}{(defaults to FALSE)}

\item{add.minor.ticks}{(defaults to TRUE)}

\item{custom.labels}{(defaults to NULL)}

\item{ticks.in.introns}{(defaults to TRUE)}

\item{position}{(defaults to "below.gene")}

\item{inverted}{(defaults to FALSE)}

\item{add.units}{(defaults to TRUE)}

\item{digits}{(defaults to 2)}

\item{shorten.labels}{(defaults to TRUE)}

\item{cDNA}{(defaults to FALSE)}

\item{cex}{(defaults to 0.6)}

\item{label.col}{(defaults to NULL)}

\item{tick.dist}{(defaults to 500)}

\item{tick.len}{(defaults to 0.2)}

\item{tick.col}{(defaults to NULL)}

\item{minor.tick.dist}{(defaults to 100)}

\item{minor.tick.len}{(defaults to 0.5)}

\item{minor.tick.col}{(defaults to NULL)}

\item{r0}{(defaults to 0)}

\item{r1}{(defaults to 1)}

\item{data.panel}{(defaults to 1)}

\item{col}{(defaults to "black")}

\item{pos}{(defaults to NULL)}

\item{...}{}
}
\value{
invisibly returns the given GenoPlot object
}
\description{
Adds a ruler with base numbers to the plot
}
\details{
This function adds a ruler to the plot marking the position of the plot elements in bases.
It can either label the ends of the exons or create a label every a certain number of bases
or both. The base numbers can be plotted any where in the genoplot as any other plot element
but it's possible to plot them just below or just above the gene structure setting
the parameter \code{position} to  \code{"below.gene"} (default) or \code{"above.gene"}.
The representation is quite customizable and it's possible to decide if ticks will be
plotted on intronic regions, if we want minor ticks between the labelled ticks,
the colors, sizes and distances for major and minor ticks, whether labels should have
units...

With the parameter \code{col} it's possible to determine the color of all elements, but
there are parameters to specify the color of individual elements (labels, ticks...) and
these take precendence over \code{col}.
}
\examples{

gp <- plotGene(exons=toGRanges(c("chr1:1-200", "chr1:300-350")))
gpAddBaseNumbers(gp)

}
\seealso{
\code{\link{plotGene}}
}
