% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpAddGeneLabel.R
\name{gpAddGeneLabel}
\alias{gpAddGeneLabel}
\title{gpAddGeneLabel}
\usage{

}
\value{
invisibly returns the given GenoPlot object
}
\description{
Plots the exons and introns structure used to create the GenoPlot
}
\details{
Given a GenoPlot object
}
\examples{


}
\seealso{
\code{\link{plotGene}}
}
