% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpAddMainTitle.R
\name{gpAddMainTitle}
\alias{gpAddMainTitle}
\title{gpAddMainTitle}
\usage{
gpAddMainTitle(genoplot, main=NULL, ...)
}
\arguments{
\item{genoplot}{a \code{GenoPlot} object returned by a call to \code{plotGene}}

\item{main}{(character) the main title of the plot}

\item{...}{any additional parameter to be passed to the text plotting. All R base graphics params are passed along.}
}
\value{
invisibly returns the given GenoPlot object
}
\description{
Plots the main title of the plot
}
\details{
Given a GenoPlot object and a character string, plot the character strings
as the main title of the plot. This function is usually automatically
called by \code{\link{plotGene}}.
}
\examples{


}
\seealso{
\code{\link{plotGene}}
}
