% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpAddBaseNumbers.R
\name{gpAddBaseNumbers}
\alias{gpAddBaseNumbers}
\title{gpAddBaseNumbers}
\usage{
gpAddMainTitle(genoplot, main=NULL, ...)
}
\arguments{
\item{genoplot}{a \code{GenoPlot} object returned by a call to \code{plotGene}}

\item{...}{any additional parameter to be passed to the text plotting. All R base graphics params are passed along.}

\item{main}{(character) the main title of the plot}
}
\value{
invisibly returns the given GenoPlot object
}
\description{
Plots the main title of the plot
}
\details{
Given a GenoPlot object and a character string, plot the character strings
as the main title of the plot. This function is usually automatically
called by \code{\link{plotGene}}.
}
\examples{

gp <- plotGene(exons=c("chr1:100-200", "chr2:300-350"))
gpAddMainTitle(gp, main="EXAMPLE", col="red", srt=30)

}
\seealso{
\code{\link{plotGene}}
}
