% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotGene}
\alias{plotGene}
\title{plotGene}
\usage{
plotGene()
}
\arguments{
\item{main}{(character) the main title of the plot}

\item{...}{any additional parameter to be passed to the text plotting. All R base graphics params are passed along.}

\item{genoplot}{a \code{GenoPlot} object returned by a call to \code{plotGene}}
}
\value{
invisibly returns the given GenoPlot object
}
\description{
Plots the main title of the plot
}
\details{

}
\examples{




}
\seealso{
\code{\link{gpAddMainTitle}}
}
